Use Tokens with FileManagerStrategy
- FileManagerStrategy now only has basic read and write functions. When
reading data you're given a token. When done sending this to CloudWatch
you can call resolve() on the token which marks that section of the file
as complete. After all sections of a file have uploaded the file is
deleted.
- FileManagerStrategy also now inherits from DataManagerStrategy so that
in the future people can build new systems for managing their offline
logs and easily hook them into the existing framework.

Signed-off-by: Miaofei <miaofei@amazon.com>